Add type checking script
Add a script for type checking code with mypy.

Note that mypy is configured to type check files with the pattern
`**/*.py`. This does not include files in the root directory. Mypy
complains if the pattern `*.py` is present in the configuration because
there do not currently exist any Python files in that directory. When we
add the `setup.py` file we should make sure to adjust the mypy files
list to include `*.py`.