Remove venvs and dist from type checked files
Remove the `venvs` and `dist` directories from the list of type checked
files by mypy. Unfortunately, mypy does not yet offer a way to exclude
files which makes this a bit more brittle than I would like. One way to
fix this would be to add a test which makes sure that the number of
files that mypy checks is the expected number of files.