Change wheel testing
Change the method for testing the wheel. Instead of running the Python
tests on it, just check that it has the same contents as the wheel built
from source.

Since the wheel build from source is tested with the pytests, this is
still checked. Also, it makes sure that the two wheels match.