Add venv usage to scripts
Add virtual environment usage to scripts that were not using them. This
adds a lot of time, but I don't think it will be bad after making venvs
reusable. We also might want to consider combining the developer and
test venvs/requirements so that only one venv is needed for the all
script.