Move wheel requirement
Move the wheel requirement to the basic requirements file. The reason
for doing this is partially because it is needed for building the
distributions, but by default is excluded from the the pip freeze
output. The other reason is that some packages use it to install (or
prefer it for installing over the legacy setup.py way.