Add standard repr name mapping
Add an argument to the standard representation for mapping parameter
names to instance attribute names. This is helpful if an initialization
argument name doesn't match the name of the instance attribute that it
is assigned to. For example if the instance attribute is private as
denoted by a leading underscore.