Change get instance attributes to return a copy
Change get instance attributes for a class to return a copy of the
dictionary of instance attribtutes. This is so that it can be modified
without changing the instance. I'm not sure if this is a deep copy, but
I'm also not sure if we would want a deep or shallow copy. I think
shallow is probably all that we want?