Add string join
Add the ability to join strings with seperatos and optionally a
different final seperator. For example, you can join the strings 'foo',
'bar', and 'baz' into 'foo, bar or baz'. The next thing for this would be
to add support for a different seperator depending on whether there are
only two, or two or more seperators (this would let you do Oxford
commas.) We might also figure out a better looping mechanism? For
example something that lets you loop through an iterable with a kernel
of current value, next value or size kernels?